move sensitivity plugins to one package
