+++ app/src/main/java/info/nightscout/androidaps/plugins/Sensitivity/SensitivityAAPSPlugin.java
@@ -1,10 +1,12 @@
-package info.nightscout.androidaps.plugins.SensitivityWeightedAverage;
+package info.nightscout.androidaps.plugins.Sensitivity;
 
 import android.support.v4.util.LongSparseArray;
 
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
+import java.util.ArrayList;
+import java.util.Arrays;
 import java.util.Date;
 import java.util.List;
 
@@ -28,24 +30,24 @@ import info.nightscout.utils.SafeParse;
  * Created by mike on 24.06.2017.
  */
 
-public class SensitivityWeightedAveragePlugin extends PluginBase implements SensitivityInterface {
-    private static Logger log = LoggerFactory.getLogger(SensitivityWeightedAveragePlugin.class);
+public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInterface {
+    private static Logger log = LoggerFactory.getLogger(SensitivityAAPSPlugin.class);
 
-    private static SensitivityWeightedAveragePlugin plugin = null;
+    static SensitivityAAPSPlugin plugin = null;
 
-    public static SensitivityWeightedAveragePlugin getPlugin() {
+    public static SensitivityAAPSPlugin getPlugin() {
         if (plugin == null)
-            plugin = new SensitivityWeightedAveragePlugin();
+            plugin = new SensitivityAAPSPlugin();
         return plugin;
     }
 
-    public SensitivityWeightedAveragePlugin() {
+    public SensitivityAAPSPlugin() {
         super(new PluginDescription()
                 .mainType(PluginType.SENSITIVITY)
-                .pluginName(R.string.sensitivityweightedaverage)
+                .pluginName(R.string.sensitivityaaps)
                 .shortName(R.string.sensitivity_shortname)
                 .preferencesId(R.xml.pref_absorption_aaps)
-                .description(R.string.description_sensitivity_weighted_average)
+                .description(R.string.description_sensitivity_aaps)
         );
     }
 
@@ -60,33 +62,30 @@ public class SensitivityWeightedAveragePlugin extends PluginBase implements Sens
         if (age.equals(MainApp.gs(R.string.key_child))) defaultHours = 4;
         int hoursForDetection = SP.getInt(R.string.key_openapsama_autosens_period, defaultHours);
 
+        Profile profile = MainApp.getConfigBuilder().getProfile();
+
+        if (profile == null) {
+            log.debug("No profile");
+            return new AutosensResult();
+        }
+
         if (autosensDataTable == null || autosensDataTable.size() < 4) {
-            if (Config.logAutosensData)
-                log.debug("No autosens data available");
+            log.debug("No autosens data available");
             return new AutosensResult();
         }
 
         AutosensData current = IobCobCalculatorPlugin.getPlugin().getAutosensData(toTime); // this is running inside lock already
         if (current == null) {
-            if (Config.logAutosensData)
-                log.debug("No autosens data available");
+            log.debug("No autosens data available");
             return new AutosensResult();
         }
 
-
-        Profile profile = MainApp.getConfigBuilder().getProfile();
-        if (profile == null) {
-            if (Config.logAutosensData)
-                log.debug("No profile available");
-            return new AutosensResult();
-        }
 
         List<CareportalEvent> siteChanges = MainApp.getDbHelper().getCareportalEventsFromTime(fromTime, CareportalEvent.SITECHANGE, true);
 
+        List<Double> deviationsArray = new ArrayList<>();
         String pastSensitivity = "";
         int index = 0;
-        LongSparseArray<Double> data = new LongSparseArray<>();
-
         while (index < autosensDataTable.size()) {
             AutosensData autosensData = autosensDataTable.valueAt(index);
 
@@ -100,14 +99,9 @@ public class SensitivityWeightedAveragePlugin extends PluginBase implements Sens
                 continue;
             }
 
-            if (autosensData.time < toTime - hoursForDetection * 60 * 60 * 1000L) {
-                index++;
-                continue;
-            }
-
             // reset deviations after site change
             if (CareportalEvent.isEvent5minBack(siteChanges, autosensData.time)) {
-                data.clear();
+                deviationsArray.clear();
                 pastSensitivity += "(SITECHANGE)";
             }
 
@@ -117,11 +111,10 @@ public class SensitivityWeightedAveragePlugin extends PluginBase implements Sens
             if (autosensData.bg < 80 && deviation > 0)
                 deviation = 0;
 
-            //data.append(autosensData.time);
-            long reverseWeight = (toTime - autosensData.time) / (5 * 60 * 1000L);
-            data.append(reverseWeight, deviation);
-            //weights += reverseWeight;
-            //weightedsum += reverseWeight * (autosensData.validDeviation ? autosensData.deviation : 0d);
+            if (autosensData.time > toTime - hoursForDetection * 60 * 60 * 1000L)
+                deviationsArray.add(deviation);
+            if (deviationsArray.size() > hoursForDetection * 60 / 5)
+                deviationsArray.remove(0);
 
 
             pastSensitivity += autosensData.pastSensitivity;
@@ -132,41 +125,26 @@ public class SensitivityWeightedAveragePlugin extends PluginBase implements Sens
             index++;
         }
 
-        if (data.size() == 0) {
-            return new AutosensResult();
-        }
-
-        double weightedsum = 0;
-        double weights = 0;
-
-        long hightestWeight = data.keyAt(data.size() - 1);
-        for (int i = 0; i < data.size(); i++) {
-            long reversedWeigth = data.keyAt(i);
-            double value = data.valueAt(i);
-            double weight = (hightestWeight - reversedWeigth) / 2;
-            weights += weight;
-            weightedsum += weight * value;
-        }
-
-        if (weights == 0) {
-            return new AutosensResult();
-        }
+        Double[] deviations = new Double[deviationsArray.size()];
+        deviations = deviationsArray.toArray(deviations);
 
         double sens = profile.getIsf();
 
         String ratioLimit = "";
-        String sensResult;
+        String sensResult = "";
 
         if (Config.logAutosensData)
             log.debug("Records: " + index + "   " + pastSensitivity);
 
-        double average = weightedsum / weights;
-        double basalOff = average * (60 / 5) / Profile.toMgdl(sens, profile.getUnits());
+        Arrays.sort(deviations);
+
+        double percentile = IobCobCalculatorPlugin.percentile(deviations, 0.50);
+        double basalOff = percentile * (60 / 5) / Profile.toMgdl(sens, profile.getUnits());
         double ratio = 1 + (basalOff / profile.getMaxDailyBasal());
 
-        if (average < 0) { // sensitive
+        if (percentile < 0) { // sensitive
             sensResult = "Excess insulin sensitivity detected";
-        } else if (average > 0) { // resistant
+        } else if (percentile > 0) { // resistant
             sensResult = "Excess insulin resistance detected";
         } else {
             sensResult = "Sensitivity normal";
@@ -181,12 +159,13 @@ public class SensitivityWeightedAveragePlugin extends PluginBase implements Sens
 
         if (ratio != rawRatio) {
             ratioLimit = "Ratio limited from " + rawRatio + " to " + ratio;
-            if (Config.logAutosensData)
-                log.debug(ratioLimit);
+            log.debug(ratioLimit);
         }
 
-        if (Config.logAutosensData)
-            log.debug("Sensitivity to: " + new Date(toTime).toLocaleString() + " weightedaverage: " + average + " ratio: " + ratio + " mealCOB: " + current.cob);
+        if (Config.logAutosensData) {
+            log.debug("Sensitivity to: " + new Date(toTime).toLocaleString() + " percentile: " + percentile + " ratio: " + ratio + " mealCOB: " + current.cob);
+            log.debug("Sensitivity to: deviations " + Arrays.toString(deviations));
+        }
 
         AutosensResult output = new AutosensResult();
         output.ratio = Round.roundTo(ratio, 0.01);
