+++ app/src/main/java/info/nightscout/androidaps/plugins/Sensitivity/SensitivityOref1Plugin.java
@@ -1,4 +1,4 @@
-package info.nightscout.androidaps.plugins.SensitivityAAPS;
+package info.nightscout.androidaps.plugins.Sensitivity;
 
 import android.support.v4.util.LongSparseArray;
 
@@ -27,40 +27,33 @@ import info.nightscout.utils.SP;
 import info.nightscout.utils.SafeParse;
 
 /**
- * Created by mike on 24.06.2017.
+ * Created by mike on 19.06.2018.
  */
 
-public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInterface {
-    private static Logger log = LoggerFactory.getLogger(SensitivityAAPSPlugin.class);
+public class SensitivityOref1Plugin extends PluginBase implements SensitivityInterface {
+    private static Logger log = LoggerFactory.getLogger(IobCobCalculatorPlugin.class);
 
-    static SensitivityAAPSPlugin plugin = null;
+    static SensitivityOref1Plugin plugin = null;
 
-    public static SensitivityAAPSPlugin getPlugin() {
+    public static SensitivityOref1Plugin getPlugin() {
         if (plugin == null)
-            plugin = new SensitivityAAPSPlugin();
+            plugin = new SensitivityOref1Plugin();
         return plugin;
     }
 
-    public SensitivityAAPSPlugin() {
+    public SensitivityOref1Plugin() {
         super(new PluginDescription()
                 .mainType(PluginType.SENSITIVITY)
-                .pluginName(R.string.sensitivityaaps)
+                .pluginName(R.string.sensitivityoref1)
                 .shortName(R.string.sensitivity_shortname)
-                .preferencesId(R.xml.pref_absorption_aaps)
-                .description(R.string.description_sensitivity_aaps)
+                .preferencesId(R.xml.pref_absorption_oref1)
+                .description(R.string.description_sensitivity_oref1)
         );
     }
 
     @Override
     public AutosensResult detectSensitivity(long fromTime, long toTime) {
         LongSparseArray<AutosensData> autosensDataTable = IobCobCalculatorPlugin.getPlugin().getAutosensDataTable();
-
-        String age = SP.getString(R.string.key_age, "");
-        int defaultHours = 24;
-        if (age.equals(MainApp.gs(R.string.key_adult))) defaultHours = 24;
-        if (age.equals(MainApp.gs(R.string.key_teenage))) defaultHours = 4;
-        if (age.equals(MainApp.gs(R.string.key_child))) defaultHours = 4;
-        int hoursForDetection = SP.getInt(R.string.key_openapsama_autosens_period, defaultHours);
 
         Profile profile = MainApp.getConfigBuilder().getProfile();
 
@@ -76,10 +69,9 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
 
         AutosensData current = IobCobCalculatorPlugin.getPlugin().getAutosensData(toTime); // this is running inside lock already
         if (current == null) {
-            log.debug("No autosens data available");
+            log.debug("No current autosens data available");
             return new AutosensResult();
         }
-
 
         List<CareportalEvent> siteChanges = MainApp.getDbHelper().getCareportalEventsFromTime(fromTime, CareportalEvent.SITECHANGE, true);
 
@@ -111,11 +103,12 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
             if (autosensData.bg < 80 && deviation > 0)
                 deviation = 0;
 
-            if (autosensData.time > toTime - hoursForDetection * 60 * 60 * 1000L)
-                deviationsArray.add(deviation);
-            if (deviationsArray.size() > hoursForDetection * 60 / 5)
-                deviationsArray.remove(0);
+            deviationsArray.add(deviation);
 
+            for (int i = 0; i < autosensData.extraDeviation.size(); i++)
+                deviationsArray.add(autosensData.extraDeviation.get(i));
+            if (deviationsArray.size() > 96)
+                deviationsArray.remove(0);
 
             pastSensitivity += autosensData.pastSensitivity;
             int secondsFromMidnight = Profile.secondsFromMidnight(autosensData.time);
@@ -125,11 +118,24 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
             index++;
         }
 
+        // when we have less than 8h worth of deviation data, add up to 90m of zero deviations
+        // this dampens any large sensitivity changes detected based on too little data, without ignoring them completely
+        log.debug("Using most recent " + deviationsArray.size() + " deviations");
+        if (deviationsArray.size() < 96) {
+            int pad = Math.round((1 - deviationsArray.size() / 96) * 18);
+            log.debug("Adding " + pad + " more zero deviations");
+            for (int d = 0; d < pad; d++) {
+                //process.stderr.write(".");
+                deviationsArray.add(0d);
+            }
+        }
+
         Double[] deviations = new Double[deviationsArray.size()];
         deviations = deviationsArray.toArray(deviations);
 
         double sens = profile.getIsf();
 
+        double ratio = 1;
         String ratioLimit = "";
         String sensResult = "";
 
@@ -138,13 +144,26 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
 
         Arrays.sort(deviations);
 
-        double percentile = IobCobCalculatorPlugin.percentile(deviations, 0.50);
-        double basalOff = percentile * (60 / 5) / Profile.toMgdl(sens, profile.getUnits());
-        double ratio = 1 + (basalOff / profile.getMaxDailyBasal());
+        for (double i = 0.9; i > 0.1; i = i - 0.01) {
+            if (IobCobCalculatorPlugin.percentile(deviations, (i + 0.01)) >= 0 && IobCobCalculatorPlugin.percentile(deviations, i) < 0) {
+                if (Config.logAutosensData)
+                    log.debug(Math.round(100 * i) + "% of non-meal deviations negative (>50% = sensitivity)");
+            }
+            if (IobCobCalculatorPlugin.percentile(deviations, (i + 0.01)) > 0 && IobCobCalculatorPlugin.percentile(deviations, i) <= 0) {
+                if (Config.logAutosensData)
+                    log.debug(Math.round(100 * i) + "% of non-meal deviations negative (>50% = resistance)");
+            }
+        }
+        double pSensitive = IobCobCalculatorPlugin.percentile(deviations, 0.50);
+        double pResistant = IobCobCalculatorPlugin.percentile(deviations, 0.50);
 
-        if (percentile < 0) { // sensitive
+        double basalOff = 0;
+
+        if (pSensitive < 0) { // sensitive
+            basalOff = pSensitive * (60 / 5) / Profile.toMgdl(sens, profile.getUnits());
             sensResult = "Excess insulin sensitivity detected";
-        } else if (percentile > 0) { // resistant
+        } else if (pResistant > 0) { // resistant
+            basalOff = pResistant * (60 / 5) / Profile.toMgdl(sens, profile.getUnits());
             sensResult = "Excess insulin resistance detected";
         } else {
             sensResult = "Sensitivity normal";
@@ -152,6 +171,8 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
 
         if (Config.logAutosensData)
             log.debug(sensResult);
+
+        ratio = 1 + (basalOff / profile.getMaxDailyBasal());
 
         double rawRatio = ratio;
         ratio = Math.max(ratio, SafeParse.stringToDouble(SP.getString(R.string.key_openapsama_autosens_min, "0.7")));
@@ -162,10 +183,8 @@ public class SensitivityAAPSPlugin extends PluginBase implements SensitivityInte
             log.debug(ratioLimit);
         }
 
-        if (Config.logAutosensData) {
-            log.debug("Sensitivity to: " + new Date(toTime).toLocaleString() + " percentile: " + percentile + " ratio: " + ratio + " mealCOB: " + current.cob);
-            log.debug("Sensitivity to: deviations " + Arrays.toString(deviations));
-        }
+        if (Config.logAutosensData)
+            log.debug("Sensitivity to: " + new Date(toTime).toLocaleString() + " ratio: " + ratio + " mealCOB: " + current.cob);
 
         AutosensResult output = new AutosensResult();
         output.ratio = Round.roundTo(ratio, 0.01);
